requirements.txt
flask==3.0.2
flask-cors==4.0.0
opencv-python==4.9.0.80
tensorflow>=2.16.0
numpy>=1.26.0
pillow>=10.0.0
transformers>=4.35.0
torch>=2.1.0
speechrecognition>=3.10.0
pydub>=0.25.1
textblob>=0.17.1
scikit-learn>=1.3.2
pandas>=2.1.0
python-dotenv>=1.0.0
werkzeug>=3.0.1
